requirements.txt
Jinja2==3.1.6
MarkupSafe==3.0.3
filelock==3.20.2
fsspec==2025.12.0
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.0
pillow==12.1.0
pip==25.3
setuptools==80.9.0
sympy==1.14.0
torch @ file:///home/tim/activity/DINO-X/rocm-wheels/torch-2.9.1%2Brocm7.1.1.lw.git351ff442-cp313-cp313-linux_x86_64.whl#sha256=d63d5938c533d3c4d6b5e6532cf3f28028c0897c780fbbe03d11792ca3e914cb
torchaudio @ file:///home/tim/activity/DINO-X/rocm-wheels/torchaudio-2.9.0%2Brocm7.1.1.gite3c6ee2b-cp313-cp313-linux_x86_64.whl#sha256=c0c7d049bd75457a5a07d96c74a93d078fb614575ccf714878a26f28fa399fde
torchvision @ file:///home/tim/activity/DINO-X/rocm-wheels/torchvision-0.24.0%2Brocm7.1.1.gitb919bd0c-cp313-cp313-linux_x86_64.whl#sha256=33f0900eb0eb9b2bcb70f7458c08f00e75d08bf85342e1d93144c7a9ea4f4fb9
triton @ file:///home/tim/activity/DINO-X/rocm-wheels/triton-3.5.1%2Brocm7.1.1.gita272dfa8-cp313-cp313-linux_x86_64.whl#sha256=46bfcf4efe3638814b7bd4611b293abe454487ad4f685d5a7da12d6e3a9c4fa0
typing_extensions==4.15.0
pydicom==3.0.1
